clean up code and create mode modules to avoid putting intelligence in Cmd interactive
