+++ setup.py
@@ -161,7 +161,7 @@ class SdistCommand(sdist):
 
 setuptools.setup(
     name="virt-scenario",
-    version="1.0.6",
+    version="2.0.0",
     author="Antoine Ginies",
     author_email="aginies@suse.com",
     description="Virt-scenario",
