+++ src/virtscenario/immutable.py
